Moved the following functions into EWSApiModule:
- recover_soft_delete_items
- create_folder
- mark_item_as_junk
    - MarkAsJunk class
